Update skr 1.3 extruder currents
This changes the SKR 1.3 run and hold current to match the config in
the Voron-2 repo and the SKR 1.4 repo. The 1A currents listed here can
lead to the extruder stepper getting hot enough to melt the extruder
body.  Based on feedback in Discord, this appears to be a problem a few
people have had.