New coal container with lid. (#516)
* Delete 1_coal_for_infill_mesh.stl

New coal container with lid.

* Delete 1_coal_with_mesh.stl

New coal container with lid.

* New coal container with lid.

A new coal container with lid to avoid mess when moving the printer. The printing instructions are still the same as before. Many thanks to Kevin Rahhal!

* Delete Lid.stl

* New coal container with lid.

A new coal container with lid to avoid mess when moving the printer. The printing instructions are still the same as before. Many thanks to Kevin Rahhal!

* New coal container with lid.

A new coal container with lid to avoid mess when moving the printer. The printing instructions are still the same as before.

* Delete VEFACH.step

* VEFACH - Voron Exhaust Filter Activated Coal + Hepa

With a new coal container with lid to avoid mess when moving the printer.