V0 Electronics Cooling Fan
I noticed my V0.1 was getting a few problems with overheating so I
decided to do something about it. I made a fan duct using a 24V 3010
blower fan and a small circuit board to control it This seems to work
quite well so far in my testing.

I am releasing this modification under GPLv3 in the hopes that it'll
help others in the future.