Add files via upload
Based on the work of "Fiction" I designed a Guide for the PTFE tube to prevent kinking. The two pieces work together and create a 90° tunnel with a diameter of 75 mm. Space needed behind the enclosure 25 mm. 

You need two heating inserts and two screws SDHS M3 8mm to mount the two parts together.