Create Keshka V2.5417
I drew up a case to hold my new BTT 7" touch screen I wanted to use with my OrangePi and Voron 2.4r2 although it will mount to any 2020 rail. Suggested mods would be a quick release for the pivot points so the screen can easily be rotated out of the way of a 270' door. The case has a tab out one side for cable strain release and a detachable foot to hold it away from the side panels of the Voron while "touching" the screen.