use native Jinja variable _bucket_pos
The previous implementation of dynamically changing bucket_pos on each
run had an extremely confusing semantic of changing the variable early
in the execution of the macro but see the effect only on the next run of
the macro.  This led to the situation that the first bucket position
after the machine started was never selected randomly, only all later
ones.  While this for sure is not a major problem of the implementation
it is extremely confusing and makes extending the macro unnecessarily
complicated.

See the confusion this created on my side at
https://klipper.discourse.group/t/conditional-changing-of-a-variable-inside-a-macro/19161

By using the native Jinja variable _bucket_pos instead, we can make the
change effective immediately, making the macro behave how you would
intuitively expect.