variable_bucket_park: select bucket to park nozzle after cleaning
The previous implementation always parked the nozzle over the left
bucket after cleaning.  While this makes sense in most cases it is
problematic when you happen to have something installed over the left
bucket, like the Klicky Probe.  If parking the nozzle over the Klicky
Probe there is a certain risk to pull it out of the docking station by
accident with the move command following the cleaning procedure.  When
we park the nozzle over the right bucket we can mitigate that risk.

The default value for variable_bucket_park is still to park over the
left bucket as in the original implementation.