variable_bucket_purge: select bucket for purging
In the original implementation, the bucket for purging was selected
randomly.  Since I wanted to park the nozzle always over the right
bucket to avoid conflicts with the Klicky Probe and as such always
oozing into that (on my 350 build) smaller right bucket I thought that
it might be useful to compensate for this by always purging to the
larger left bucket.

Now you can select the bucket you want to do the purging.  With the
default setting of -1 it will still randomly choose the bucket as
before.